Giving the possibility of overriding default port
Hello Guys, 
A small change that can give possibilities of set port for Expo DevTools.
Any others part of expo modules gives the possibility of overriding of the default port e.g. by setting `"manifestPort": 19790` in file `.exprc` or by setting `"packagerOpts": {"port": 19797}` in file `app.json`,
this is the last package that has no such possibility.

The setting of own ports is very helpful when someone must work with more than one app in the project. The issue that we have is that apps on stop/start mostly change ports and that makes a lot of chaos on connected devices.